refactor(e2e)/move e2e tests to root directory
